# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GreenLink & Sunrise Agrovet Retail Operations OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List/search products
      parameters:
        - name: shopId
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
          description: Search query (fuzzy)
        - name: category
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: string
          description: Filter to low stock only
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Soft-delete product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /products/{productId}/restock:
    post:
      operationId: restockProduct
      tags: [products]
      summary: Restock product (adds stock + creates inventory movement)
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestockInput"
      responses:
        "200":
          description: Restocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}/price-history:
    get:
      operationId: getProductPriceHistory
      tags: [products]
      summary: Get product price history
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Price history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceHistory"
  /products/bulk-import:
    post:
      operationId: bulkImportProducts
      tags: [import]
      summary: Bulk import products from parsed Excel data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"

  # ─── PRODUCT ALIASES ─────────────────────────────────────
  /products/{productId}/aliases:
    get:
      operationId: listProductAliases
      tags: [aliases]
      summary: List aliases for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Aliases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductAlias"
    post:
      operationId: createProductAlias
      tags: [aliases]
      summary: Add alias to product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductAliasInput"
      responses:
        "201":
          description: Alias created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductAlias"
  /aliases/{aliasId}:
    delete:
      operationId: deleteProductAlias
      tags: [aliases]
      summary: Delete alias
      parameters:
        - name: aliasId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── SALES ───────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
        shopId:
          type: string
        canonicalName:
          type: string
        normalizedName:
          type: string
        sku:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        unit:
          type: string
        purchasePrice:
          type: ["number", "null"]
        sellingPrice:
          type: ["number", "null"]
        profitMargin:
          type: ["number", "null"]
        stockQty:
          type: integer
        alertQty:
          type: integer
        tokens:
          type: ["array", "null"]
          items:
            type: string
        size:
          type: ["string", "null"]
        isActive:
          type: boolean
        lastSoldAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, shopId, canonicalName, normalizedName, unit, stockQty, alertQty, isActive, createdAt, updatedAt]
    ProductList:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
      required: [products, total]
    ProductInput:
      type: object
      properties:
        shopId:
          type: string
        canonicalName:
          type: string
        sku:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        unit:
          type: string
        purchasePrice:
          type: ["number", "null"]
        sellingPrice:
          type: ["number", "null"]
        stockQty:
          type: integer
        alertQty:
          type: integer
        size:
          type: ["string", "null"]
      required: [shopId, canonicalName]
    ProductUpdate:
      type: object
      properties:
        canonicalName:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        purchasePrice:
          type: ["number", "null"]
        sellingPrice:
          type: ["number", "null"]
        stockQty:
          type: ["integer", "null"]
        alertQty:
          type: ["integer", "null"]
        size:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    RestockInput:
      type: object
      properties:
        qty:
          type: integer
        newPurchasePrice:
          type: ["number", "null"]
        newSellingPrice:
          type: ["number", "null"]
        performedBy:
          type: string
      required: [qty]
    PriceHistory:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        oldPurchasePrice:
          type: ["number", "null"]
        newPurchasePrice:
          type: ["number", "null"]
        oldSellingPrice:
          type: ["number", "null"]
        newSellingPrice:
          type: ["number", "null"]
        pctChange:
          type: ["number", "null"]
        changedAt:
          type: string
        changedBy:
          type: ["string", "null"]
      required: [id, productId, changedAt]
    BulkImportInput:
      type: object
      properties:
        shopId:
          type: string
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductInput"
        deduplicateStrategy:
          type: string
          enum: [skip, merge, overwrite]
          description: What to do with duplicate products
      required: [shopId, products]
    BulkImportResult:
      type: object
      properties:
        created:
          type: integer
        skipped:
          type: integer
        merged:
          type: integer
        errors:
          type: array
          items:
            type: string
      required: [created, skipped, merged, errors]

    # ─── ALIASES ──────────────────────────────────────────
    ProductAlias:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        alias:
          type: string
        confidence:
          type: ["number", "null"]
        autoGenerated:
          type: boolean
        createdAt:
          type: string
      required: [id, productId, alias, autoGenerated, createdAt]
    ProductAliasInput:
      type: object
      properties:
        alias:
          type: string
        confidence:
          type: ["number", "null"]
      required: [alias]

    # ─── SALES ────────────────────────────────────────────
...